+++ src/core/occ/request.js
@@ -5,6 +5,8 @@ var winston = require('winston');
 var request = require('request');
 var fs = require('fs-extra');
 var config = require('../config');
+var _Configs = require('../configs');
+var CreateWorkset = require("../worksets/api/create");
 
 /**
  * Mount the config object.
@@ -184,7 +186,7 @@ function tryToRequest(config, token, fileToken, maxAttempts, callback) {
     });
   };
 
-  var proceedWithRequest = function (err, callback) {
+  var proceedWithRequest = function (config, err, callback) {
     if(err) {
       return callback(err);
     }
@@ -224,9 +226,9 @@ function tryToRequest(config, token, fileToken, maxAttempts, callback) {
     },
     function(callback) {
       if(tokenIsValid) {
-        isPublishRunning.call(self, config, token, fileToken, proceedWithRequest, callback);
+        isPublishRunning.call(self, config, token, fileToken, setWorkset.bind(self, proceedWithRequest), callback);
       } else {
-        proceedWithRequest(null, callback);
+        proceedWithRequest(config, null, callback);
       }
     },
     callback
@@ -246,7 +248,7 @@ function isPublishRunning(config, token, fileToken, next, callback) {
   }
 
   if(notAffectByPublishMethods.includes(method)) {
-    return next(null, callback);
+    return next(config, null, callback);
   }
 
   const publishStatusConfig = mountConfig(this._endpoint, 'publish?lastPublished=true');
@@ -257,7 +259,7 @@ function isPublishRunning(config, token, fileToken, next, callback) {
     }
 
     if(!body.publishRunning) {
-      return next(null, callback);
+      return next(config, null, callback);
     }
 
     winston.info('There is a publish in progress, waiting for this process to be finished...');
@@ -267,6 +269,111 @@ function isPublishRunning(config, token, fileToken, next, callback) {
       isPublishRunning.apply(this, arguments);
     }, PUBLISH_CHECK_DELAY);
   });
+}
+
+/**
+ * Sets workset name in the configs if it's not present
+ *
+ * @param   {String}    workset   The workset name
+ * @param   {Function}  callback  Async module callback function
+ *
+ */
+function createWorksetConfigIfNotAvailable(workset, callback) {
+  if(config.workset) {
+    return callback();
+  }
+
+  var occToolsConfigs = new _Configs();
+  occToolsConfigs.setWorkset({ workset }, callback);
+}
+
+/**
+ * Creates the workset and assign the workset
+ * id to the configs and to all requests
+ *
+ * @param   {Object}  requestData  The request configs object
+ * @param   {Function}  callback   The Async module callback
+ *
+ */
+var worksetError = false;
+function setWorkset(next, requestData, error, callback) {
+  if(error) {
+    return callback(error);
+  }
+
+  var byPassApis = ['login', 'worksets', 'publish'];
+  var url = requestData.url || request.api;
+
+  var isByPassApi = byPassApis.find(function (item) {
+    return url.includes(item);
+  });
+
+  var setWorksetIdHeaders = function() {
+    if(config.worksetId) {
+      requestData.headers = {
+        ...requestData.headers,
+        'X-CC-Workset': config.worksetId
+      }
+    }
+  };
+
+  // if workwsetId already available
+  // set it in the headers request
+  setWorksetIdHeaders();
+
+  // If it's the allowed endpoint, just dontinue the process
+  if(isByPassApi) {
+    return next(requestData, null, callback);
+  }
+
+  // If the workset is not set yet and there is no error while
+  // setting the workset, try to create and get the workset id
+  if(!config.worksetId && !worksetError) {
+    var createWorkset = new CreateWorkset(this);
+    var username = config.credentials.username;
+    var worksetName = config.workset || username.substring(0, username.indexOf('@'));
+
+    createWorkset.new(worksetName)
+      .then(function (data) {
+        var foundWorksets = data.worksets;
+
+        if(config.worksetId) {
+          return next(requestData, null, callback);
+        }
+
+        if(foundWorksets && !foundWorksets.length) {
+          winston.warn("Problem while fetching/creating the workset...");
+          worksetError = true;
+          return next(requestData, null, callback);
+        }
+
+        createWorksetConfigIfNotAvailable(worksetName, function (error) {
+          if(error) {
+            return callback(error);
+          }
+
+          config.worksetId = foundWorksets[0].repositoryId;
+          config.workset = worksetName;
+
+          winston.info('');
+          winston.info(`Adding changes to the workset "${config.workset}"...`);
+          winston.info('');
+
+          setWorksetIdHeaders();
+          next(requestData, null, callback);
+        });
+      })
+      .catch(function (error) {
+        worksetError = true;
+        winston.warn("Error while fetching/creating the workset...", error.message);
+        next(requestData, null, callback);
+      });
+
+      return;
+  }
+
+  // Other regular cases no related to the workset creation
+  return next(requestData, null, callback);
 }
 
 module.exports = function(rawOpts, callback) {